- migrated reobfJar task over to ACP-Gradle. See the latest commit (cd231c8) in ACP-Gradle's repository. - merged the "modifyJar" task into the default Gradle "jar" task.
Initially I wanted both the jar construction stuff and the reobfuscation to be in the plugin, but I had way too much trouble getting the jar constructing integrated into the ACPPlugin class properly. This was due to a number of reasons, primarily dealing with the Copy task (and its inheritors) and how it works in Java (Java has no closures, unlike Groovy, which made it far more difficult to get things configured the way I wanted). I'll chalk up most of it to inexperience on my part. Still frustrating nonetheless.